restructure test to match src
